refactor guards and strategies
