+++ apps/api/src/auth/guards/jwt-auth.guard.ts
@@ -5,8 +5,8 @@ import {
 } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
 import { AuthGuard } from '@nestjs/passport';
-import { IS_PUBLIC_KEY } from './is-public';
-import { AuthService } from './auth.service';
+import { AuthService } from '../auth.service';
+import { IS_PUBLIC_KEY } from '../is-public';
 
 @Injectable()
 export class JwtAuthGuard extends AuthGuard('jwt') {
